sửa lại api room và entity

+++ BackEnd/HotelBooking/src/main/java/com/example/HotelBoking/DTO/ReviewDTO.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 public class ReviewDTO {
     private Long id;
     private Long userId;
-    private Long hotelId;
+    private Integer hotelId;
     private Integer rating;
     private String comment;
     private LocalDateTime createdAt;
