+++ BackEnd/HotelBooking/src/main/java/com/example/HotelBoking/Entity/Hotel.java
@@ -9,36 +9,27 @@ import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "hotels")
-
 public class Hotel {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String name;
-
     private String address;
-
     @Column(name = "city")
     private String city;
-
-
     @Column(columnDefinition = "TEXT")
     private String description;
-
     @Column(name = "star_rating")
     private Integer starRating;
-
     @Column(name = "created_at")
     private LocalDateTime createdAt;
-
     private BigDecimal latitude;
     private BigDecimal longitude;
 
     public Hotel(){}
 
-    public Hotel(Long id, String name, String address, String city, String description, Integer starRating, LocalDateTime createdAt) {
+    public Hotel(Integer id, String name, String address, String city, String description, Integer starRating, LocalDateTime createdAt) {
         this.id = id;
         this.name = name;
         this.address = address;
@@ -48,59 +39,18 @@ public class Hotel {
         this.createdAt = createdAt;
     }
 
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public Integer getStarRating() {
-        return starRating;
-    }
-
-    public void setStarRating(Integer starRating) {
-        this.starRating = starRating;
-    }
-
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(LocalDateTime createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
+    public String getName() { return name; }
+    public void setName(String name) { this.name = name; }
+    public String getAddress() { return address; }
+    public void setAddress(String address) { this.address = address; }
+    public String getDescription() { return description; }
+    public void setDescription(String description) { this.description = description; }
+    public Integer getStarRating() { return starRating; }
+    public void setStarRating(Integer starRating) { this.starRating = starRating; }
+    public LocalDateTime getCreatedAt() { return createdAt; }
+    public void setCreatedAt(LocalDateTime createdAt) { this.createdAt = createdAt; }
+    public String getCity() { return city; }
+    public void setCity(String city) { this.city = city; }
 }
