+++ BackEnd/HotelBooking/src/main/java/com/example/HotelBoking/Repository/HotelRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface HotelRepository extends JpaRepository<Hotel, Long> {
+public interface HotelRepository extends JpaRepository<Hotel, Integer> {
     List<Hotel> findByCity(String city);
 }
